requirements.txt
tuya-iot-python-sdk==0.4.1
Pillow==10.0.0
